requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1
redis==5.0.1
asyncpg==0.29.0

# Message Queue
kafka-python==2.0.2
aiokafka==0.10.0

# ML/AI
scikit-learn==1.4.0
xgboost==2.0.3
numpy==1.26.3
pandas==2.1.4
tensorflow==2.15.0

# LLM & AI
# openai==1.10.0  # Optional: Only needed if using OpenAI LLM (not AWS Bedrock)
anthropic==0.8.1
langchain==0.1.4
chromadb==0.4.22

# Voice Services
twilio==8.11.1
# openai-whisper==20231117  # Optional: Only needed if using OpenAI STT (not AWS Transcribe)
# elevenlabs==0.2.27  # Optional: Only needed if using ElevenLabs TTS (not AWS Polly)

# Geospatial
geoalchemy2==0.14.3
shapely==2.0.2

# HTTP & API
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==42.0.0

# AWS
boto3==1.34.34
botocore==1.34.34

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
hypothesis==6.98.1
faker==22.2.0

# Development
black==24.1.1
ruff==0.1.15
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
ipython==8.20.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tenacity==8.2.3
